+++ scripts/erc-4626/migrate-vault-strategy-tags.py
@@ -45,6 +45,7 @@ from eth_defi.enzyme.tags import get_strategy_tags as get_enzyme_strategy_tags
 from eth_defi.erc_4626.core import ERC4262VaultDetection, ERC4626Feature, get_vault_protocol_name
 from eth_defi.erc_4626.vault_protocol.aave.tags import get_strategy_tags as get_aave_strategy_tags
 from eth_defi.erc_4626.vault_protocol.atoma.tags import STRATEGY_TAGS as ATOMA_STRATEGY_TAGS
+from eth_defi.erc_4626.vault_protocol.axis.tags import STRATEGY_TAGS as AXIS_STRATEGY_TAGS
 from eth_defi.erc_4626.vault_protocol.centrifuge.tags import STRATEGY_TAGS as CENTRIFUGE_STRATEGY_TAGS
 from eth_defi.erc_4626.vault_protocol.ethena.tags import STRATEGY_TAGS as ETHENA_STRATEGY_TAGS
 from eth_defi.erc_4626.vault_protocol.euler.tags import get_strategy_tags as get_euler_strategy_tags
@@ -233,6 +234,7 @@ EVM_STRATEGY_TAG_RESOLVERS: dict[ERC4626Feature, StrategyTagResolver] = {
     ERC4626Feature.upshift_like: partial(lookup_strategy_tags, UPSHIFT_STRATEGY_TAGS),
     ERC4626Feature.centrifuge_like: partial(lookup_strategy_tags, CENTRIFUGE_STRATEGY_TAGS),
     ERC4626Feature.ethena_like: partial(lookup_strategy_tags, ETHENA_STRATEGY_TAGS),
+    ERC4626Feature.axis_like: partial(lookup_strategy_tags, AXIS_STRATEGY_TAGS),
     ERC4626Feature.yieldnest_like: partial(lookup_strategy_tags, YIELDNEST_STRATEGY_TAGS),
     ERC4626Feature.enzyme_blue_like: get_enzyme_strategy_tags,
 }